requirements.txt
# Core blockchain interaction
web3==6.11.3
requests==2.31.0
aiohttp==3.8.6
asyncio-throttle==1.0.2

# Database connections
influxdb-client==1.38.0
psycopg2-binary==2.9.7

# Data processing and analysis
pandas==2.1.1
numpy==1.25.2
pyarrow==13.0.0

# Ethereum/EVM specific tools
eth-account==0.9.0
eth-utils==2.2.1
eth-abi==4.2.1
hexbytes==0.3.1

# Crypto and hashing
cryptography==41.0.5
pycryptodome==3.19.0

# Configuration and utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.6.0
tqdm==4.66.1

# Logging and monitoring
structlog==23.1.0
prometheus-client==0.17.1

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1

# Jupyter for analytics
jupyter==1.0.0
matplotlib==3.8.0
seaborn==0.12.2
plotly==5.17.0

# Performance optimization
ujson==5.8.0
orjson==3.9.7
multiprocessing-logging==0.3.4

# Optional: Machine learning for advanced analytics
scikit-learn==1.3.1

# Web service for monitoring dashboard
aiohttp==3.8.6
aiohttp-cors==0.7.0
